Support JSON-formatted string as messages
Using BarbequeClient for enqueuing a message, it might be Array or Hash
because it is posted as JSON format with `Content-Type: application/json` header.
However in some cases, e.g. posted as query parameter style or enqueued
via SNS subscriptions, the message comes as JSON-formatted string. So
this patch make a message which is String parse as JSON.