Change S3 log format
I'm planning to change the timing when message body is stored to S3
because it's currently hard to recover from unexpected errors during job
executions. If the message was stored to S3 before starting job
executions, it might be possible to retry the enqueued job even if the
Barbeque worker host goes down or the database server goes down while
running the job.

In order to change the timing, I'd like to split the log file because
message body can be stored before job execution but stdout and stderr
cannot be stored before job execution.